mesa xserver xkbcomp git update 12 Mar 2014
xserver          commit 81a4952d3dcab9ca3a1ee399c773f5ac352036c4
xkbcomp          commit d4e02a09258063c6d024c3ccd42d6b22212e6e18
mesa             commit 2c886eba7820f7e3e965dd16b786b0a412d8ef11

+++ xorg-server/test/hashtabletest.c
@@ -12,7 +12,7 @@ static void
 print_xid(void* ptr, void* v)
 {
     XID *x = v;
-    printf("%ld", *x);
+    printf("%ld", (long)(*x));
 }
 
 static void
